Factor out hostport splitting for use by both catalog and chirp.
